+++ src/components/UI/screenSizes.ts
@@ -1,0 +1,11 @@
+const sizes = {
+    wide: '1275px',
+    medium: '925px',
+    narrow: '600px'
+}
+
+export const screenWidth = {
+    wideWidth: `min-width: ${sizes.wide}`,
+    mediumWidth: `min-width: ${sizes.medium}`,
+    narrowWidth: `min-width: ${sizes.narrow}`
+}